refactor(server): fix 3 code quality issues in applications module
- Move generateCandidates logic (incl. candidateSchema) to StaffService;
  CommonStaffService now delegates to staffService.generateCandidates()
- Remove duplicate openrouter instance and createOpenAI import from
  CommonStaffService; single source of truth lives in StaffService
- Replace redundant CommonStaffResult interface with a re-export of
  StaffBotResult from staff.service.ts
- Export StaffService from ApplicationsModule so other modules can inject it
